rtl8723bs: fix broken jiffies calcs
The existing code was not dealing with jiffies wraparound,
that was resulting in the card sometimes being left permanently
on or off via IPS.

+++ drivers/net/wireless/rtl8723bs/include/drv_types.h
@@ -589,7 +589,7 @@ struct dvobj_priv
 	unsigned char	oper_channel; //saved channel info when call set_channel_bw
 	unsigned char	oper_bwmode;
 	unsigned char	oper_ch_offset;//PRIME_CHNL_OFFSET
-	u32 on_oper_ch_time;
+	unsigned long on_oper_ch_time;
 
 	//extend to support mulitu interface
 	//padapters[IFACE_ID0] == if1
